Add files via upload
The python file shows the following task 1: 
The JSON file contains 1000 records of products. Each product has the following 4 attributes 
Subcategory
Title
Price
Popularity 
Task:
Fetch the JSON file programmatically and store the data in the data structure of your choice.
Display the data in the presentation of your choice with Title, Price ordered based on the descending popularity.